Expanded NeuralNetwork::FeedForward
Previous was just stubby. I've tried to make extensive use of the STL
Algorithm library and Boost's slicing. It was a PITA trying to use
multi_array iterators until I read the headers/documentation and found
that ArrayType::reference is the value iterated over in the first row.

Added a LinearActivationFunction.